RF - renamed scil_tractogram_filter_anatomically.py
