+++ scripts/scil_tractogram_filter_by_anatomy.py
